+++ src/Products/urban/profiles/extra/config_default_values.py
@@ -1180,5 +1180,36 @@ default_values = {
         'streets': [
             'City',
         ],
+
+        'karst_constraints': [
+            'UrbanVocabularyTerm',
+            {'id': "no", 'title': u"Sans"},
+            {'id': "low", 'title': u"Faible"},
+            {'id': "moderate", 'title': u"Modéré"},
+            {'id': "high", 'title': u"Fort"},
+        ],
+
+        'concentrated_runoff_s_risk': [
+            'UrbanVocabularyTerm',
+            {'id': "low", 'title': u"Faible, bassin versant afférent entre 1 et 9 Ha"},
+            {'id': "medium", 'title': u"Moyen, bassin versant afférent entre 9 et 18 Ha"},
+            {'id': "high", 'title': u"Elevé, bassin versant afférent de plus de 18 Ha"},
+        ],
+
+        'seveso_site': [
+            'UrbanVocabularyTerm',
+            {'id': "exxonmobil", 'title': u"EXXONMOBIL PETROLEUM & CHEMICAL TOURNAI (Seuil bas)"},
+            {'id': "nyco", 'title': u"NYCO - STPC TOURNAI (Seuil bas)"},
+            {'id': "orica", 'title': u"ORICA - TITA NOBEL BELGIQUE S.A. TOURNAI (Seuil haut)"},
+            {'id': "univar", 'title': u"UNIVAR BELGIUM - QUARON TOURNAI (Seuil Haut)"},
+            {'id': "rynders", 'title': u"RYNDERS PETROLEUM BELGIUM ANTOING (seuil bas)"},
+            {'id': "transmat", 'title': u"TRANSMAT DEPOT ANTOING (seuil bas)"},
+        ],
+
+        'natura_2000': [
+            'UrbanVocabularyTerm',
+            {'id': "rumillies", 'title': u"Périmètre NATURA 2000 du Bas Follet (Rumillies)"},
+            {'id': "vaulx", 'title': u"Périmètre NATURA 2000 de la GRANDE MER (Vaulx)"},
+        ],
     },
 }