+++ docs/CHANGES.rst
@@ -4,7 +4,71 @@ Changes
 =======
 
 
-1.11.0 (unreleased)
+1.11.1 (unreleased)
+-------------------
+- improve removeItems utils
+  [fngaha]
+
+- Refactor rename natura2000 field because of conflict name in thee
+  [fngaha]
+
+- Refactor getFirstAdministrativeSfolderManager to getFirstGradeIdSfolderManager
+  The goal is to use one method to get any ids
+  [fngaha]
+
+- Add generic SEVESO optional fields
+  [fngaha]
+
+- Fix concentratedRunoffSRisk and details optional fields
+  [fngaha]
+
+- Add getFirstAdministrativeSfolderManager method
+  [fngaha]
+
+- Add removeItems utils and listSolicitOpinionsTo method
+  [fngaha]
+
+- Add getFirstDeposit and _getFirstEvent method
+  [fngaha]
+
+- remove the character 'à' in the address signaletic
+  [fngaha]
+
+- use RichWidget for 'missingPartsDetails', 'roadMissingPartsDetails', 'locationMissingPartsDetails'
+  [fngaha]
+
+- Fix local workday's method"
+  [fngaha]
+
+- Add a workday method from collective.delaycalculator
+  refactor getUrbanEvents by adding UrbanEventOpinionRequest
+  rename getUrbanEventOpinionRequest to getUrbanEvent
+  rename containsUrbanEventOpinionRequest to containsUrbanEvent
+  [fngaha]
+
+- Add methods
+  getUrbanEventOpinionRequests
+  getUrbanEventOpinionRequest
+  containsUrbanEventOpinionRequest
+  [fngaha]
+
+- Update askFD() method
+  [fngaha]
+
+- Add generic Natura2000 optional fields
+  [fngaha]
+
+- Fix codec in getMultipleClaimantsCSV (when use a claimant contat)
+  [fngaha]
+
+- Add generic concentratedRunoffSRisk and details optional fields
+  [fngaha]
+
+- Add generic karstConstraint field and details optional fields
+  [fngaha]
+
+
+1.11.0 (2015-10-01)
 -------------------
 
 - Nothing changed yet.
@@ -29,7 +93,7 @@ Changes
 
 - Add environment licence class one.
 
-- Bug fix: config folder are not allowed anymore to be selected as values 
+- Bug fix: config folder are not allowed anymore to be selected as values
   for the field 'additionalLegalConditions'.
 
 
@@ -76,7 +140,7 @@ Changes
 - Add a richer context for python expression in urbanEvent default text.
 
 - Factorise all licence views through a new generic, extendable and customisable view (#6942).
-  The fields display order is now given by the licence class schemata and thus this order 
+  The fields display order is now given by the licence class schemata and thus this order
   is always consistent between the edit form and the view form.